refactored to use latest preact
