+++ package.json
@@ -1,67 +1,53 @@
 {
   "private": true,
   "name": "architecture-viewer",
-  "version": "1.0.0",
-  "license": "Apache-2.0",
+  "version": "0.0.0",
+  "license": "MIT",
   "scripts": {
-    "start": "if-env NODE_ENV=production && npm run -s serve || npm run -s dev",
-    "build": "preact build --no-prerender",
-    "serve": "preact build --no-prerender && preact serve",
+    "build": "preact build",
+    "serve": "sirv build --port 8080 --cors --single",
     "dev": "preact watch",
     "lint": "eslint src",
-    "test": "jest ./tests",
-    "deploy": "gulp deploy",
-    "license": "license-checker --csv --production > licenses.csv"
+    "test": "jest"
   },
   "eslintConfig": {
-    "extends": "eslint-config-synacor"
+    "extends": "preact",
+    "ignorePatterns": [
+      "build/"
+    ]
   },
-  "eslintIgnore": [
-    "build/*"
-  ],
   "devDependencies": {
-    "eslint": "^4.9.0",
-    "eslint-config-synacor": "^2.0.2",
-    "gulp": "^3.9.1",
-    "gulp-gh-pages": "^0.5.4",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-preact-pure": "^2.0.0",
+    "eslint": "^6.0.1",
+    "eslint-config-preact": "^1.1.0",
     "identity-obj-proxy": "^3.0.0",
-    "if-env": "^1.0.0",
-    "jest": "^21.2.1",
-    "license-checker": "^20.2.0",
-    "preact-cli": "^2.1.0",
-    "preact-render-spy": "^1.2.1"
+    "jest": "^24.9.0",
+    "jest-preset-preact": "^1.0.0",
+    "jest-raw-loader": "^1.0.1",
+    "preact-cli": "^3.0.0",
+    "preact-render-spy": "^1.2.1",
+    "sirv-cli": "1.0.3"
   },
   "dependencies": {
-    "ajv": "^6.5.3",
-    "cytoscape": "^3.2.16",
-    "cytoscape-dagre": "^2.2.1",
-    "cytoscape-popper": "^1.0.2",
-    "dagre": "^0.8.2",
-    "preact": "^8.2.6",
-    "preact-compat": "^3.17.0",
-    "preact-router": "^2.5.7",
+    "ajv": "^6.12.3",
+    "cytoscape": "^3.15.2",
+    "cytoscape-dagre": "^2.2.2",
+    "cytoscape-popper": "^1.0.7",
+    "dagre": "^0.8.5",
+    "preact": "^10.3.2",
+    "preact-render-to-string": "^5.1.4",
+    "preact-router": "^3.2.1",
     "tippy.js": "^2.5.4"
   },
   "jest": {
-    "verbose": true,
-    "setupFiles": [],
-    "testURL": "http://localhost:8080",
-    "moduleFileExtensions": [
-      "js",
-      "jsx"
+    "preset": "jest-preset-preact",
+    "setupFiles": [
+      "<rootDir>/tests/__mocks__/browserMocks.js",
+      "<rootDir>/tests/__mocks__/setupTests.js"
     ],
-    "moduleDirectories": [
-      "node_modules"
-    ],
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/src/tests/__mocks__/fileMock.js",
-      "\\.(css|less|scss)$": "identity-obj-proxy",
-      "^./style$": "identity-obj-proxy",
-      "^preact$": "<rootDir>/node_modules/preact/dist/preact.min.js",
-      "^react$": "preact-compat",
-      "^react-dom$": "preact-compat",
-      "^create-react-class$": "preact-compat/lib/create-react-class",
-      "^react-addons-css-transition-group$": "preact-css-transition-group"
+    "transform": {
+      "\\.adoc$": "jest-raw-loader"
     }
   }
 }
