--- src/assets/parsejson.js
@@ -1,30 +1,0 @@
-/*
- * Copyright 2017 Capital One Services, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-let fs = require('fs');
-
-let obj = JSON.parse(fs.readFileSync('grid_data.json', 'utf8'));
-
-obj = obj.map(o => {
-    o.position = undefined;
-    return o;
-})
-
-// console.log(obj);
-
-let res = JSON.stringify(obj);
-
-fs.writeFileSync('nopos_data.json', res, 'utf8')
