+++ tests/testfiles/doubled_names.adoc
