+++ tests/testfiles/participant_decl.adoc
