+++ tests/testfiles/steps_group_mixed.adoc
